requirements.txt
torch
torchvision
torchaudio
torchtext
tensorboard
pyyaml
tqdm
numpy
pandas
datasets
Pillow
matplotlib
scikit-learn